chore: refactor builder workflows
